Fix(inputitem): add wai-aria suppport for the money type InputItem (#2538)
* Fix(inputitem): add wai-aria suppport for the money type InputItem

The money type InputItem is simulated with the div tag, rather than the origin
input tag. To support wai-aria needs a little more work, for the div tag, by adding
role of `textbox` make div perform like input in aria mode

For the CustomKeyboard, it will lead to wrong focus when not adding aria
supoort for backspace and keyboard canceling

Fixed: #2419, #2418

* Fix(inputitem): remove extra useless custom keyboard instance

In the money type of InputItem, the CustomKeyboard will create as many times as the InputItem is
used.The inner variable `customNumberKeyboard` will point to the latest custom keyboard instance
which can avoid the repeat appearing animation. But when the InputItem is render async, since the
`customNumberKeyboard` points to the latest custom keyboard instance, the operation for previous
custom keyboard instance such as add hidden class  will be invalid

To Solve this problem, remove the useless instances of custom keyboard
everytime one InputItem is being blured

Fixed: #2475